Move AppControlPanel.vue to components/AppControlPanel/
