+++ app/components/SlidingPanel/components/AppControlPanel/AppControlPanel.vue
